CAN-IO: Split PWM from Servo modules
